+++ packages/ok-components/src/ts/table-column/testing/ts-table-column-breakpoint.pageobject.ts
@@ -1,7 +1,7 @@
 import type { TablePageObject } from '@ni/nimble-components/dist/esm/table/testing/table.pageobject';
 import type { TableRecord } from '@ni/nimble-components/dist/esm/table/types';
-import { BreakpointState } from '../types';
-import { TsTableColumnBreakpointCellView } from '../cell-view';
+import { BreakpointState } from '../breakpoint/types';
+import { TsTableColumnBreakpointCellView } from '../breakpoint/cell-view';
 
 /**
  * Page object for ts-table-column-breakpoint tests.
